refactor: useWave 훅 분리
